requirements.txt
streamlit
tensorflow
keras
Pillow
numpy
matplotlib
torch
torchvision
torchaudio
openimages